+++ src/prec/graph-reducer.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const N3            = require('n3');
-const DStar         = require('../dataset/index.js');
+const { default: DStar }   = require('../dataset/index');
 const namespace     = require('@rdfjs/namespace');
 
 const Context  = require("./context-loader.js");
