+++ test/index.js
@@ -7,5 +7,5 @@ require("./dataset/DatasetCore.test.js")({
     quad: N3.DataFactory.quad,
     literal: N3.DataFactory.literal,
     blankNode: N3.DataFactory.blankNode,
-    dataset: t => new (require('../src/dataset/index.js'))(t)
+    dataset: t => new (require('../src/dataset/index').default)(t)
 });
