+++ test/prec_impl/prec-c-template-deducing.test.js
@@ -4,7 +4,7 @@ const EdgeRules = require('../../src/prec/rules-for-edges');
 const PropertyRules = require('../../src/prec/rules-for-properties');
 const N3 = require("n3");
 const { readRawTemplate } = require('../../src/prec/context-loader');
-const DStar = require('../../src/dataset/index.js');
+const { default: DStar } = require('../../src/dataset/index');
 
 const prefixes =
 `
